+++ app/ehrbase-rest/src/api/definition/dispatch.rs
@@ -47,7 +47,7 @@ async fn run(
         "definition_template_adl2_upload" => template_adl2::upload(&state, &parts).await,
         "definition_template_adl2_get" => template_adl2::get(&state, &parts).await,
         "definition_template_adl2_example_get" => template_adl2::example_get(&parts),
-        "definition_template_adl2_version_get" => template_adl2::version_get(&parts),
+        "definition_template_adl2_version_get" => template_adl2::version_get(&state, &parts).await,
         "definition_query_list" => stored_query::list(&state, &parts).await,
         "definition_query_list_all" => stored_query::list_all(&state, &parts).await,
         "definition_query_store.yaml" => stored_query::store(&state, &parts).await,